requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy==2.2.4
ollama==0.4.7
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pytesseract==0.3.13
sniffio==1.3.1
typing_extensions==4.12.2
